+++ src/components/ui/Button.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Slot } from './SimpleSlot';
+import { Slot } from './Slot';
 import type { VariantProps } from 'class-variance-authority';
 
 import { cn } from '@/lib/utils';
